Move processing-report semantics into compatibility
Amp-Thread-ID: https://ampcode.com/threads/T-019ff0ab-0530-75ba-a946-8c2e9e9544ca
